requirements.txt
airflow-clickhouse-plugin
airflow-provider-kafka
apache-airflow-providers-google
apache-airflow-providers-http
apache-airflow-providers-slack
apache-airflow-providers-postgres
apache-airflow-providers-slack[http]
apache-airflow[sentry]
pyarrow==10.0.1
pydantic==1.10.2
cron-validator==1.0.6
git+https://github.com/mullakhmetov/ethereum-etl.git@fix-contract-type-detection
airflow-dbt==0.4.0
dbt-core==1.4.6
dbt-postgres==1.4.6
dbt-clickhouse==1.4.0
retry
pandas
clickhouse_driver==0.2.5
ethereum-etl
plyvel
annoy==1.17.2
scikit-learn==1.2.2